analyzer/requirements.txt
kafka-python==2.1.5
google-generativeai>=0.8.0
pydantic>=2.0
python-dotenv>=1.0
colorama>=0.4.6
